Cluster Depth Tests

Compute the Cluster- multiple comparisons procedure.

compute_clusterdepth_array(distribution, threshold, graph, alternative, border)

Arguments

distribution
An 3d array representing the null distribution of multiple signal. The first dimension is the permutations, the second the samples, the third is the channels.

threshold
A numeric specifying the height of the threshold to compute the clusters.

graph
A igraph object representing the adjacency of the channels (used for the outputs only).

alternative
A character string indicating the alternative hypothesis. Either "greater", "less" or "two.sided".

border
A character string indicating the method to handles the border. Either "reverse", reversing the computation of the cluster-depth for clusters at the border of the time frame, or "ignore".

Value

A list containing an igraph and a data.frame, with the results for each sample, channel.

References

Frossard, J., & Renaud, O. (2021). The Cluster Depth Tests: Toward Point-Wise Strong Control of the Family-Wise Error Rate in Massively Univariate Tests with Application to M/EEG. arXiv preprint arXiv:2105.07514.

See also