Step-down version of the max T multiple comparisons procedure

Compute the step-down version of the max T multiple comparisons procedure.

compute_stepdownmaxT_array(distribution, alternative, graph)

Arguments

distribution

An 3d array representing the null distribution of multiple signal. The first dimension is the permutations, the second the samples, the third is the channels.

alternative

A character string indicating the alternative hypothesis. Either "greater", "less" or "two.sided".

graph

A igraph object representing the adjacency of the channels (used for the outputs only).

Value

A list containing an igraph and a data.frame, with the results for each sample, channel.

See also