Summary method for brainperm object

List per effect of the results of the brainperm object.

# S3 method for brainperm
summary(
  object,
  multcomp = NULL,
  table_type = "cluster",
  alternative = "two.sided",
  ...
)

Arguments

object
a brainperm object.

multcomp
a character string indicating the multiple comparison procedure to display. Only relevant if several multiple comparison procedures are computed in the brainperm object.

table_type
a character string indicating the type of table. The default value, table_type = "cluster", displays results by clusters. If table_type = "full", the table shows the results by individual tests.

alternative
a character string indicating the alternative hypothesis for t-test. The default value if "two.sided". "less" and "greater" are also available.

...
other arguments pass to print.

Value

a list of tables containing the results of the test for each effect.

See also